Quick Sort - 1.2
Fixed the sorting algorithm, it is working perfectly.
There were a few gross errors in the code, especially
in the partition() function. The rest of the program
seems to be working fine, I will be implementing the
randomized quick sort version as soon as I reach home.
Finally got some work done!